renamed folder and added test bucket support
