+++ data-mining/data-process-fnc/readme.md
@@ -1,0 +1,6 @@
+## How it works
+When the raw data are uploaded, we will split them in chunks for 70k per region. 
+This is the maximum number of requests we can make a day to riot's api
+
+Cloud function triggered by GCS upload?
+https://cloud.google.com/functions/docs/calling/storage
