+++ data-mining/data-process-fnc/requirements.txt
@@ -1,0 +1,2 @@
+pandas==1.1.4
+google-cloud-storage==1.32.0
