--- data-mining/data-processing/readme.md
@@ -1,2 +1,0 @@
-## How it works
-After the data has been split, we will use a similar technique as the raw-data processor to gather the match data.
