+++ data-mining/match-data/readme.md
@@ -1,6 +1,2 @@
 ## How it works
-When the raw data are uploaded, we will split them in chunks for 70k per region. 
-This is the maximum number of requests we can make a day to riot's api
-
-Cloud function triggered by GCS upload?
-https://cloud.google.com/functions/docs/calling/storage
+After the data has been split, we will use a similar technique as the raw-data processor to gather the match data.
